feat: Add function to extract top 10 colors from images
This commit adds a new function `get_colors` to the `get_top10_colors.py` file. The function takes a list of images and masks, and generates PNGs with histograms of the top 10 most common colors in each group. It also saves the histograms as PNG files. This function will be useful for analyzing color distributions in different image groups.